pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-cloud-removal"
version = "0.1.0"
description = "Benchmark generative models and multi-temporal fusion for satellite cloud removal."
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pillow>=10.0",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "scipy>=1.10",
  "tqdm>=4.66",
  "torch>=2.1",
  "torchvision>=0.16"
]

[project.optional-dependencies]
geo = ["earthengine-api>=0.1.390", "rasterio>=1.3"]
dev = ["pytest>=7.4"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
